css/css-content/content-animation.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>content animation</title>
<link rel="help" href="https://drafts.csswg.org/css-content/#content-property">
<meta name="test" content="box-shadow supports animation">
<link rel="author" href="mailto:graouts@apple.com" title="Antoine Quint">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

.target::after {
  content: none;
}

.target.initial::after {
  content: "initial";
}

@keyframes content-animation {
  from { content: "from" }
  to   { content: "to" }
}

.target.animated::after {
  animation: content-animation 1s paused linear forwards;
}

</style>
<body>
<script>

const contentAnimationTest = (className, expectedInitialValue, title) => {
    test(() => {
        const target = document.body.appendChild(document.createElement("div"));
        target.classList.add("target");
        if (className)
            target.classList.add(className);

        const style = getComputedStyle(target, "::after");

        assert_equals(style.content, expectedInitialValue, "Before the animation is applied.");

        target.classList.add("animated");
        const animation = target.getAnimations({ subtree: true })[0];

        const testContentAtTime = (time, expected) => {
          animation.currentTime = time;
          assert_equals(style.content, `"${expected}"`, `Check the animated value at time = ${time}ms`);
        };

        testContentAtTime(0, 'from');
        testContentAtTime(499, 'from');
        testContentAtTime(500, 'to');
        testContentAtTime(999, 'to');
        testContentAtTime(1000, 'to');

        target.remove();
    }, title);
}

contentAnimationTest('', '', 'The content property can be animated with a discrete animation type when there is no initial content value.');
contentAnimationTest('initial', '"initial"', 'The content property can be animated with a discrete animation type when there is an initial content value.');

</script>
</body>
